refactor(core): implement locking and atomic dongle acquisition
- remove `dongle_manager` and its global mutex to eliminate threading
bottlenecks.
- add individual `access_mutex` and `access_cond` to each `t_dongle`.
- implement atomic "all-or-nothing" dongle acquisition to prevent
hold-and-wait deadlocks.
- add `start_mutex` in workspace to perfectly synchronize coder threads
at T=0.
- change `heapq` priority type from `int` to `size_t` to prevent
overflow.
- fix `heapq_peek` to return the full node (`t_heapq_data *`) instead of
raw data.
- update workspace and coder initializations to reflect the new
architecture.
- clean up Makefile and update all unit tests to match the new behavior.

+++ coders/src/heapq/heapq_enqueue.c
@@ -6,7 +6,7 @@
 /*   By: nlallema <nlallema@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/25 18:06:39 by nlallema          #+#    #+#             */
-/*   Updated: 2026/03/28 17:59:38 by nlallema         ###   ########lyon.fr   */
+/*   Updated: 2026/03/29 23:31:03 by nlallema         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-static t_heapq_data	*_heapq_create_data(void *data, int priority)
+static t_heapq_data	*_heapq_create_data(void *data, size_t priority)
 {
 	t_heapq_data	*heapq_data;
 
@@ -77,7 +77,7 @@ static void	_heapq_heapify_up(t_heapq *heapq)
 	}
 }
 
-t_errcode	heapq_enqueue(t_heapq *heapq, void *data, int priority)
+t_errcode	heapq_enqueue(t_heapq *heapq, void *data, size_t priority)
 {
 	int				errcode;
 	t_heapq_data	*heapq_data;
